fix subscription panels
